update packages; fix vulnerabilites and deprecations
