+++ cfg/CollectionsPlugin.js
