+++ cfg/DateFiltersPlugin.js
