+++ cfg/RandPlugin.js
