+++ cfg/StringFiltersPlugin.js
