+++ cfg/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+  CollectionsPlugin: require("./CollectionsPlugin"),
+  DateFiltersPlugin: require("./DateFiltersPlugin"),
+  UrlFiltersPlugin: require("./UrlFiltersPlugin"),
+  ImagePlugin: require("./ImagePlugin"),
+  RandPlugin: require("./RandPlugin"),
+  SassPlugin: require("./SassPlugin"),
+  StringFiltersPlugin: require("./StringFiltersPlugin"),
+  RemoveTrailingSlashPlugin: require("./RemoveTrailingSlashPlugin"),
+}
