+++ package.json
@@ -19,7 +19,6 @@
     "@11ty/eleventy-img": "^4.0.2",
     "@shopify/prettier-plugin-liquid": "^1.5.2",
     "autoprefixer": "^10.4.14",
-    "eleventy-plugin-postcss": "^1.0.4",
     "postcss": "^8.4.35",
     "prettier": "^3.3.3",
     "sass": "^1.79.4"
