+++ src/_includes/about.scss
@@ -1,3 +1,5 @@
+@use "shared" as *;
+
 .About {
   &-title {
     font-size: $font-size4;
