+++ src/_includes/events.scss
@@ -1,3 +1,5 @@
+@use "shared" as *;
+
 $list: null;
 
 .Events {
