+++ src/_includes/landing.scss
@@ -1,3 +1,5 @@
+@use "shared" as *;
+
 .Landing {
   display: flex;
   justify-content: center;
