+++ src/_includes/shared.scss
@@ -1,7 +1,6 @@
-@import "shared/reset";
-@import "shared/media";
-@import "shared/fonts";
-@import "shared/sizes";
-@import "shared/colors";
-@import "shared/mixins";
-@import "under-construction/under-construction";
+@forward "shared/reset";
+@forward "shared/media";
+@forward "shared/fonts";
+@forward "shared/sizes";
+@forward "shared/colors";
+@forward "shared/mixins";
