+++ src/site.scss
@@ -1,11 +1,12 @@
 /* -- includes --*/
-@import "_includes/shared";
-@import "_includes/nav";
-@import "_includes/landing";
-@import "_includes/events";
-@import "_includes/event-detail";
-@import "_includes/about";
-@import "_includes/merch";
+@use "_includes/shared" as *;
+@use "_includes/nav";
+@use "_includes/landing";
+@use "_includes/events";
+@use "_includes/event-detail";
+@use "_includes/about";
+@use "_includes/merch";
+@use "_includes/under-construction";
 
 /* -- resets (more) -- */
 * {
